feat: connect with swift api handler
